+++ lib/experimental/Information/Avatars/PersonAvatar/index.stories.tsx
