+++ lib/experimental/Information/Avatars/PersonAvatar/index.tsx
