+++ lib/experimental/Information/Avatars/exports.ts
@@ -1,3 +1,3 @@
 export * from "./CompanyAvatar"
 export * from "./TeamAvatar"
-export * from "./UserAvatar"
+export * from "./PersonAvatar"
